Split big `Session` class into multiple files Added support of timeout on `Session.RunTransaction` methods
